refactor: only move to pop_common the needed functions
