+++ crates/pop-common/src/lib.rs
@@ -15,7 +15,7 @@ pub use git::{Git, GitHub, Release};
 pub use helpers::{get_project_name_from_path, prefix_with_current_dir_if_needed, replace_in_file};
 pub use manifest::{add_crate_to_workspace, find_workspace_toml};
 pub use metadata::format_type;
-pub use signer::{create_signer, parse_account};
+pub use signer::create_signer;
 pub use templates::extractor::extract_template_files;
 // External exports
 pub use subxt::{Config, PolkadotConfig as DefaultConfig};
