+++ crates/pop-contracts/src/call.rs
@@ -5,7 +5,7 @@ use crate::{
 	utils::{
 		get_manifest_path,
 		metadata::{process_function_args, FunctionType},
-		parse_balance,
+		parse_account, parse_balance,
 	},
 };
 use anyhow::Context;
@@ -15,7 +15,7 @@ use contract_extrinsics::{
 	ExtrinsicOptsBuilder, TokenMetadata,
 };
 use ink_env::{DefaultEnvironment, Environment};
-use pop_common::{create_signer, parse_account, Config, DefaultConfig, Keypair};
+use pop_common::{create_signer, Config, DefaultConfig, Keypair};
 use sp_weights::Weight;
 use std::path::PathBuf;
 use url::Url;
